only print prompt info when connected to a terminal
This requires a new package: "code.google.com/p/go.crypto/ssh/terminal"

Because go's dependency handling is retarded, I'm not sure if this is
the correct thing to do. the upstream go.crypto repo has no release
versions, so the only way to get is is to download it from hg and build
the latest version. Or some arbitrary commit that isn't the latest
version, whatever. Go's build system can automate this if you have
GOPATH set, and run "go get", but I am thouroughly convinced that this
is the wrong way to go about dependency management.